+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * NOTE: admin audience changed. Migrate admin tools to use new one.
 
 ## `20221027t112600-all`
  * Bumped runtimeVersion to `2022.10.27`.